BackupShipping: prevent supplier from removing too many items
The suppliers handed to this class all have a supply-method that also
deletes the item it returns. Due to this behavior, one additional item was
always removed without being used. By getting and therefore removing the
item in the prepareNext method instead of the getNext method,
unintentionally removing too many items is prevented.