QSI: Fixed available size calculation
Prior this patch, the reported available size included the
oversubscriptionRatio (as a multiplicative factor), which led
to situations where available size was much larger than the
capacity.

In order to fix this, the new version is now able to consider
oversubscriptionRatios on storage pool level as well as on
storage pool definition level.
By default thick storage pools use a 1.0, thin pools 20.0
and diskless as well as the EBS_INIT storage pool an infinite
oversubscription ratio (be careful when trying to
multiply / divide by that :) )
Also, every "next_spawn_result" entry includes the actual
OversubscriptionRatio for that given storage pool entry.

The new version now still includes the oversubscriptionRatio
in the resulting max volume size, but not in the available size
nor in the capacity. That means, available size and capacity can be
compared against each other, while max volume size might exceed
both (which is more or less the core idea of overprovisioning).