BackupShipping: make inc shippings work with queue
To figure out whether or not an inc shipment needs to be queued, the
snapDfn it should be based on (prevSnap) needs to be decided on.
This used to happen based on a property that was only set when a
shipment finished shipping. But since the newest possible prevSnap can
sill be in the queue or currently shipping, a second property is needed.
This new prop always has the newest snapshot name stored, whereas the
old prop stores the snapshot name of the last finished backup.
These two are used the following way: When a new backup needs to be made,
the newer prop will be looked at to decide the prevSnap. When a shipping
can be started but needs to find a new prevSnap after being queued, it
will try to use the snap from the older prop as a base snap.

Please note that there is no fallback-snapDfn in case either of the snaps
stored in the props has been deleted. In a case like this, a full
backup will be made instead.