Mkfs: Add optional user and group
If eihter one is set, the other will get a default value.
If only user is set, group will be set to the user's current group.
If only group is set, user will be set to 'nobody'.

Both properties can either be the user-/group-name or their ID.
If a name is given, linstor checks the ID using `id -u <name>` or
`id -g <name>` respectively.