Autoplacer: Ignore evacuating / evicting nodes
The first step of the autoplacer is to filter out as many storage
pools as soon as possible. Autoplace settings as replicas-on-same
and -different can already be applied here since the nodes of
existing resources already fix / occupy some of the properties.
i.e. with a `replicas-on-different zone` already deployed resources
could already occupy `zone A` and `zone B` for example. In this case
nodes can already be filtered out if they have either `A` or `B`
as a value of the `Aux/zone` property.

However, this filtering has to be skipped for nodes that are being
evacuated or evicted, to allow evacuation to a new node within the
same zone.

Example:
nodeA - zone A
nodeB - zone A
nodeC - zone B

resources are initially spawned on nodeA and nodeC due to
`--replicas-on-different zone`. Evacuating or evicting nodeA should
allow the autplacer to choose nodeB as a new home for the resource.