NodeCon: Fix nodeCon only existing on one side
Prior this patch the NodeSatelliteFactory did not register the created
(in most cases remote-) node in the nodesMap, which lead to situations
where a node-connection appeared to be only registered in the local node
but not in both nodes.

Another small issue that was fixed regarding node connections is that
deleting a /Path/* property is now possible. However, currently the /Path/
namespace is more or less whitelisted (if the target NetInterface exists).
The problem is that usually there are two /Path/... properties required for
the path-feature to work properly. Currently the user is free to delete one
of them, add a third, or do whatever they want.