DB,Layer: Added abstract base class for LayerDbDrivers
The core idea is the following: All layers are required to provide an
implementation of at least AbsLayerRscDataDbDriver. The other abstract
base classes (for volume, rscDfn and vlmDfn) are optional / nullable.

The goal of this approach is to get away from the "one driver managing
many tables" concept, which is currently implemented by some layer
specific drivers (at least by DrbdLayer).

In a later patch, the DatabaseLoader will be modifed to pre-fetch all
optional data for layer-specific resources, volumes, rscDfn and vlmDfn
and will afterwards go through the result set of LayerResourceIds and
call the layer-specific AbsLayerRscDataDbDriver's load method to create
the layer specific rscData. That load method is responsible to create
the final rscData with correct rscDfnData, vlmDfnData and vlmData
(whatever from those are needed by the given layer).

That means, that although one class is now only responsible for one
(or none, in case of StorageLayer) database tables, it is the
Layer*RscDbDriver's responsibility to also access the cached data
from its Layer*RscDfn-, Layer*VlmDfn- and/or Layer*VlmDbDriver.