Db,Storage: Reworked storage db drivers
The interesting part in this rework is that there is no
LayerStorageResource database table. But that does not meant that this
driver has nothing to load, but more that this driver does not add anything
to its parent LayerResourceId. The LayerStorageRscDbDriver takes simply
the data from the LayerResourceId and creates a StorageRscData instance
from it, which can be used later unlike the dummy loading data from the
LayerResourceIdDbDriver.