Drbd,DbLoader: Fix database loading issue
If a resource was in toggle-disk state (regardless if adding or deleting)
during a controller shutdown/crash, the "ensureRscDataExists" method
performed a check on DRBD level whether or not a child rscData should
exist. False would mean that this DRBD resource is diskless and its
layer-child skips all layers between DRBD and STORAGE, where it gets a
DisklessData child.

That leads to problems when there were layers already loaded between
DRBD and STORAGE. For example a drbd,luks,storage resource, that still
had its disk (i.e. within a "disk removing" state), loaded from the
database this layer tree of drbd->luks->storage, but the
"ensureRscDataExists" method claims that it should be drbd->storage due to
"i am diskless" state. That caused a ClassCastException when the
RscStorageLayerHelper gets a LuksRscData instance.

This patch causes the check whether or not the resource is diskless to
also consider DISK_REMOVING and DISK_REMOVE_REQUESTED flags as not (yet)
being diskless, and thus making sure that the "ensureRscDataExists" method
expects the drbd->luks->storage tree instead of the diskless drbd->storage
tree.