DrbdLayer: Skip children in DRBD_DELETE state
In addition to c84c2bcdcdc33fe4a712249920393c8e91875bbd, there might be
an issue that if something goes wrong during a delete, the user tries
to help Linstor by manually deleting DRBD and the other layers and
(somewhat rightfully) expects Linstor to detect this new situation and
perform a noop. The problem is that with only the DRBD_DELETE flag being
set, all child-layers are told that they should still exist, so Linstor
would (after such a "helping intervention") still recreate all devices
beneath DRBD - which could re-trigger the original issues the admin solved
by hand.

This patch skips processing of child layers if only the DRBD_DELETE flag
is set, but not (yet) the DELETE flag.