LinstorScope: ensure enter and exit are executed correctly
scope.enter always needs to be called before entering a try, to
make sure that there is something to close when calling scope.exit.

Similarly, scope.exit always needs to be called inside of a finally,
to make sure the scope does not stay open.

Some classes that did not adhere to these rules have been amended.