NodeCon: Fixed compareTo
Prior this patch, the nodeCon.compareTo method took the wrong targetNode's
name in the second part (i.e. it compared this.source + this.target with
other.source + this.target.

The effect of this bug was that not all nodeConnections were loaded
from the database. From a 4 node setup with all 6 node connections
only A->B, B->C and C->D were loaded. A->C, A->D, B->D gone missing
due to the incorrect compareTo method.