BackupShipping: ensure clearing a node queue works correctly
When a node queue is cleared, all snapshots that were queued only
on that node need to be re-queued. During this process it needs to
be ensured that they do not land on the node they were just removed
from.
In theory it would be best if it was possible to get a new prevSnap
in case the current prevSnap cannot be used (e.g. it was deleted
while the snapshot was queued). However, that would require a
different approach for s3 and l2l, and would be especially
complicated for l2l, since there communication with the target
cluster is necessary to find a valid prevSnap, which requires the
use of a flux-loop.