NodeCon: Fix NPE on stlt startup
When the controller sends a FullSync to the satellites, that FullSync
only includes all node specific information for the node of the target
satellite, not its "other satellites".
That means, when sending a FullSync to NodeA, it will contain for example
node connection info from NodeA -> NodeB (whereas the NodeConnection itself
is bidirectional, so NodeA will properly link that NodeConnection as
NodeA <-> NodeB). However, the "other nodes" described in the FullSync will
*not* contain node connections at all, mostly for two reasons: 1) to
prevent including unnecessary nodes (i.e. if NodeA has nothing to do with
NodeC, but only a node conneciton NodeB <-> NodeC exists, NodeA does not
need to know about NodeC). 2) to prevent endless recursion where NodeA
includes all node-connections, it should also include NodeB's
node-connection which involves NodeA and all of it's node-connection, etc..

Since we now know that NodeB will have an empty nodeConnection list,
NodeA will properly create itself, NodeB (since it is needed by the
nodeConnection) as well as the nodeConnection itself. Great. The problem
this patch addresses is that NodeB (still within the execution of
NodeA's FullSync) runs the same logic, but finds its nodeConnection
(to NodeA) but the Pojo does not contain nodeConnections, so NodeB drops
the nodeConnection to NodeA. It got even worse since NodeB's "otherNode"
(i.e. NodeA) was also deleted since it had no resources and no
nodeConnections with the current node in common.

This patch fixes the issue such that only for the (actual) local node
nodeConnections are merged, since all "other nodes" will simply not include
any nodeConnections at all.