MkAvail: Fix incorrect deletion of TIE_BREAKER flags
This patch fixes a bug where a resource that was already in a
DRBD_DISKLESS state (regardless if TIE_BREAKER was set or not)
and a "resource make-available" was executed on that resource.
The problem was, that only deleting Resource.Flags.TIE_BREAKER
removes all 3 bits: TIE_BREAKER, DRBD_DISKLESS and DISKLESS.
That will cause an NPE on the satellites, since the resource itself
(with the missing DRBD_DISKLESS flag) looks like a diskful resource
but when trying to check its metadata the StorageRsc's DisklessData
returns "null" as devicePath (which is expected).

The fix is simply to delete the TIE_BREAKER and restore the
DRBD_DISKLESS flag if necessary.