SuspendMgr: Add new suspend-io manager class
This new class manages suspending and resuming IO for all layers.
For this, the layers are now told by the new suspend manager when
to suspend/resume an entire rscData (with all nested vlmData
objects).

Since we have a few special cases (like WriteCache or BCache),
this is done in two steps: First the root rscData are suepended.
This should be extended in the future that all given rscData
are suspended concurrently (multi-threaded) since in case
WriteCache or BCache are the topmost layers, those layers also need
to flush their cache properly.
In this first step, the non-root layers are ignored (even the two
special cases in setups like DRBD,WriteCache,Storage the WriteCache
is ignored in the first step). The reason for this is to simply
keep the topmost suspend executions as brief as possible so that
all participating resources are suspended as closely together as
possible.
In the second steps, all child-rscData that got ignored in the
first step are given the chance to run some non-root-suspend action
so that WriteCache and BCache can flush their content after
their parent rscData already properly suspended IO.