K8s,Etcd: Synchronize commit and rollback
Since we had to implement rollback mechanism for K8s and ETCD, we
also need to make sure that no two transactions race with other.

This could be the case when for example the satellite is responding
to an API call such that the controller sets a property, while a
user is running an unrelated command. These two operations do not
need to share any write locks, so locking itself does not always
prevent such concurrent transactions.