LINSTOR For Windows: special case verify-alg and cram-hmac-alg settings
Since WinDRBD currently only supports crc32c as 'crypto' alg
we had to patch ConfFileBuilder not to output algorithms
that WinDRBD does not understand. Usually in a mixed
Windows/Linux environment crc32c is ok, but sha1 (which
is still hardcoded) is not. This patch also considers
diskless nodes (consider two diskful Linux and one Windows
tie breaker, if we configure let's say sha1 on all nodes
the resource on the Windows side will not start).

Also the unit test for the conf file builder has been
relaxed in order to tolerate missing cram-hmac-alg and
shared secret fields. There is also a migration that
adds the 'windrbd' algorithm to the list of allowed
auto crypto algorithms.