Rsc,del: Fix tiebreaker-takeover issue
Given the scenario where we have a resource deployed on 3 nodes as diskful
resources and afterwards we configure the nodes to be in 2 zones (i.e
one node in zone A and the other two nodes in zone B) and configure
the resource-group with --replicas-on-different zone. This placement
becomes "invalid", but Linstor will not complain.

When deleting one of the zone B nodes, that deleted resource will be taken
over as a tiebreaker. Until now, taking over consisted of two steps:
1) toggle disk and 2) set tiebreaker flag. At the end of step 1) the
autohelpers were executed, which led to a check if a tiebreaker is required.
Since the just deleted resource is currently only in diskless state
(TIE_BREAKER flag is not yet set), but also in an invalid zone, the
auto-tiebreaker logic does not count that resource at all. Therefore finds
only 2 diskful + 0 diskless resource -> create a new tiebreaking resource
on a 4th node.

Once that is done, the original flux continues with step 2), enabling the
TIE_BREAKER flag on the overtaken resource, resulting in 2 diskful and
2 tiebreaking resources.

This patch fixes this problem by simply telling the toggle disk handler to
either set the DRBD_DISKLESS (default) or the TIE_BREAKER flag on the
resource, making the two steps of the takeover atomic. As a result the
auto-tiebreaker logic will find the tiebreaking resource at the end of
the toggle disk, skipping the "we need a tiebreaker on the 4th node" step.