Transactions: prevent AccessToDeletedDataException in transMgr
Since transactions save a HashSet of all the dirty objects in
case of a rollback, and the equals-method calls checkDeleted(),
there is the possibility for the following scenario:
Two objects get deleted during the same transaction. If those
objects have a hash-collision, when adding the second object
an AccessToDeletedDataException will occur due to the needed
equals-check.
To prevent this from happening a ThreadLocal variable is introduced
that deactivates the checkDeleted() call during registering an
object in the transMgr (and also during rollbacks, just to be sure).