Clone: skip cloning of unnecessary devices
This patch will only allow the data device as well as an optional
external metadata device for DRBD to be cloned. Other special
devices for i.e. WriteCache, Cache, etc.. are not cloned but newly
created once the cloning is finished.

WriteCache and BCache will flush their content to the data device
when their IO is suspended (which is done before taking a snapshot
and before cloning), so there is no need to actually clone the
more or less empty caching device.