RscReady: Fix expected online node ids
The issue was that deployedResources is a set of resources that
were just created inthe current API call, which does not include
already existing resources. Therefor if only one resource is created
the onlineNodeIds only contains that one entry, which gets removed
again later (temporarily) since a resource should not wait for a
connection:Connected state to itself.

This patch now properly includes all nodeIds from nodes thate are
online and diskful.