K8s: Fix rollback-replace
If a transaction was partially applied but before Linstor could remove the
rollback instance (which would have completed Linstor's transaction), the
data from the rollback instance could not be restored properly.
The problem is that the already updated instances in K8s already have a new
UID as the same instances backed up in the rollback instance. Replacing
the K8s data with the data from the rollback instance runs into a UID
mismatch problem.

In addition to 868db0bcfce7206bd636dccc277a0e724ea6ebe6, this patch also
removes the UID which disables the UID check.