Db,Props: Reworked PropsDbDriver
Two things to note here:
* PropsContainers can no longer use a removeAll(instanceName) since
that requires database deletion with partial primary keys. In SQL
that is obviously no problem, but in ETCD deleting entries by
the second primary-key (i.e. "/LINSTOR/$table/*:mypk2/*" is not
possible (* are only placeholders for non-existing wildcards).
* PropsContainer methods no longer rollback themself. This was
broken without recognition, since the rollback has an assert statement
that it only wants to be called from a TxMgr.