BackupShipping: add new receiveDoneRequest to l2l-shipping
When an l2l-shipment finishes but has an incremental backup already
queued, it is possible that the src-cluster tries to start sending
this inc backup before the dst-cluster has finished receiving the
previous backup. To ensure that the src-cluster waits long enough,
the dst-cluster now tells the src-cluster when it is done receiving
and only then does the src-cluster look if there are queued shippings
available to start.