NodeDb: Merged ETCD special delete into generic DbDriver
ETCD has a requirement that one ETCD key only occurs once in an
ETCD transaction. Usually Linstor deletes an entire object by
recursively delete (or in ETCD-speak using a range delete operation)
for the entire /LINTOR/<table>/<primaryKeys>/* namespace.

In case of a "linstor node delete" however, it might happen that
in the same Linstor transaction a node's DELETE flag gets set
as well as the node gets deleted. That triggers a recursive ETCD
delete for the node, but also a PUT request on the node's flag.
To prevent this issue, the dedicated NodeETCDDriver existed, since
it delete all entries one by one instead of using the recursive
delete.

This patch introduces the EtcdEngine.disableRecursiveDelete method
for this purpose.