DbLoader: use Set instead of List while loading
The DatabaseLoader first loads all objects, including resources and snapshots
and only afterwards loads all the layerData. The loadLayerData returned a list
of either resources or snapshots for that layerData was loaded. Returning a
list here implies that the same resource was included multiple times in the
list if it had multiple layer data loaded (like DRBD as well as storage).

This is now changed to a Set, simply to remove duplicates.