Db,luks: Changed encryptedPassword format in SQL
The old luks driver stored the encrypted password with .setBytes
wheres the SQL column type was VARCHAR2. The new driver relies on
the SQL type and tries to read/write to the column using .setString
which leads to compatibility issues.

Since the K8s and ETCD driver already Base64 encode the encrypted
password before storing them, this migration not only converts
the stored bytes[] into a String, but also Base64 encodes them
so that all 3 drivers behave similar.