Tests: Fixing unit tests
Apart from the simply missing bindings, the ObjectProtectionFactory
is now reverted to use the controller version instead of the
satellite version.

The main problem with the satellite version is that it seeds itself
a SatelliteTransMgr to initialize the objProtFactory. However, the
tests mostly use the actual SQLEngine, which requires a
Provider<TransactionMgrSql>, which the test will not seed since
it used the satellite version of the objProtFactory which does
seed the TransactionMgr, but not the TransactionMgrSql (since it has
no reason to).