ETCD,StltRemote: Changed delimiter
SatelliteRemote stored their name as ".stlt:[...]" to avoid
naming conflicts with names allowed for the user. However, since
ETCD in general uses ":" as primary-key delimiter, this was a problem
when parsing multiple PKs, as for example entries for the ACL
that stores the object protection path (that is composed as
"/remote/<remoteName>" and roles and other PKs. Since the remote
name included a ":", the split happened at the wrong place, causing
the ACL db driver to parse the second part of the remote name
as a role, which obviously failed.

Same applies for ".ebs:[...]" remotes for the very same reasons.