ETCD: Disable ranged delete requests
In some cases Linstor triggers a "recreate" event. That means that the same
ETCD keys are deleted as well as recreated in the same transaction.
ETCD does not allow multiple modifications of the same key in the same tx.
Although Linstor already has a deduplicator for exactly this purpse, ranged-
delete requests were still a problem, which just became more significat with
the recent DB-rework.
The problem is that deleting "/key/*" as well as creating new "/key/whatever"
entries cannot be detected and handled properly from the deduplicator, since
the deduplicator has no information about what keys are covered by the *. If
in the above case there was a second "/key/somethingelse" entry, the correct
behavior of such the recreate-event would have been to override "/key/whatever"
and to actually delete "/key/somethingelse". But as mentioned, the deduplicator
simply does not have the necessary information for this tx-translation.

By disabling ranged deletes, the above scenario would not have issued a delete
of "/key/*", but rather two separate deletes, one for "/key/whatever" and one
for "/key/somethingelse". If "/key/whatever" gets recreated, the deduplicator
will omit the delete for that key and only include the winning put operation
for the key into the resulting ETCD-tx. The delete request for
"/key/somethingelse" would be left untouched, resulting in the desired behavior.