Db: Fix FSM name
FreeSpaceManager used to have a ":" in the name if the user has not
specified anything. The user is not allowed to input a ":", so that
was used as a way to prevent naming conflicts.

However, since the database drivers now all use the same parser
for primary keys, the ETCD engine found in these cases 3 PKs instead
of the expected 2 for ObjProt and ACL entries.

Since dfb112c3aa60cd73c1e85d179a88553ce3db8315 already deleted the
objProt from FSM but missed to clean up the database, we can now
safely just remove the orphan DB entries.

To avoid these issues, we also change the delimiter from ":" to ";"
which also requires an update to every storage pool's FSM-name.