DB: Added basic export support
DbEngines are asked for a list of RawParameters for each table. Those rawParams
are converted with GenCrdCurrent.rawParamToSpec(..) into a serializable format.

Although the actual call to the drivers export method is missing, the ALL_TABLES
in GeneratedDatabaseTables are already reordered based on the SQL foreign-key
dependency tree, to ensure an order of ALL_TABLES in which the tables can be
populated without risking missing foreign-key references.