SQL: Added import db support
Basic procedure is to migrate from scratch up to the version specified in the
exported db.json, truncate all default data since the db.json also contains
those and finally run the insert statements for the data.

The import procedure must not depend or use any of the "current" database
components. That includes the current drivers as well as current version of
generated classes (including GeneratedDatabaseTables.ALL_TABLE). The reason for
this restriction is simply that we probably are importing a version that is
older (and maybe different structurally) than the current version.

Since insert statements needs to be done in a specific order regarding foreign
keys, and the importer is not allowed to rely on current database relations, the
importer must rely on a correct order of the db.json.