BackupShipping: prevent access to deleted Schedule
Previously when a schedule was deleted only the configs in the
scheduleLookupMap got deleted as well. This led to a scenario
where a config with a deleted schedule could be in the
activeShippings-set and therefore trigger an
accessToDeletedDataException.

Now all configs that use the schedule that will be deleted are
removed from all sets, preventing this exception.
The same is true for deleting remotes and rscDfns, although it is
not certain whether those could have triggered a similar exception.

This also fixes a ConcurrentModificationException that could happen
when more than one config of the same schedule was in the lookupMap
when the schedule should be deleted. Since we were iterating over
the same set we wanted to clean up this exception happened, but
now we create a new set to iterate over so this is no longer possible.