K8s,Crd: limit number of entries in rollback resource
Kubernetes has a limit on the maximum size of a resource: It's around 1M, which
is directly inherited from the backing ETCD store of the K8s API. This becomes
an issue in our Rollback resource: since we store a lot of other resources
inside, we may try to apply a rollback resource bigger than this limit.

The fix is to split the rollbacks into multiple resources, each limited to
a specific number of items within. We can then simply list all rollbacks
to get the complete rollback again. Since we roughly know the size of
our own resources, we can use a reasonably low limit like 100 updates
per rollback to ensure we stay below the size limit with impacting
performance in the common case.