ETCD,import: Convert Date to long
Right now only SpaceTracking uses Dates as a type. The Specific
drivers already store the Dates as long, but the importer should
not rely on driver specific implementations. For now the importer
however still converts Dates to longs.

In the long run we should migrate SpaceTracking to also just store
long instead of Dates.