RestApi: report skipDisk back to client
This change adds a map to the rsc-json called effective-properties. This
map is based on the structure of PrioProps.getConflicingMap().
With this it is possible to see where each prop is set and what will happen
if the prop is deleted on a specific object. For now, only
DrbdOptions/SkipDisk is reported in that way (mainly because there is no
consistent prioProps-ordering), but other props can be added to this map as
needed.