Db,Node: init nodes with offlinePeers
The NodeFactory already called .setOfflinePeer for new nodes. If the same node
gets created by the database (upon startup) and this method is not called and
an IOException occurs during the first connect-attempt, the node's peer instance
stays null, possibly causing NPEs when accessed by other components that want
to check if the node is online or not (like the autoplacer does).