ETCD,export: Skip empty tables
Previously an empty table was still iterated (based on its columns,
not its (missing) data), which led to an exception that a column
was unexpectedly null. However that error is usually only a problem
if only one part of a data-entry is null (which it is not supposed
to be). If the entire (or in this case all) entry is missing, we
are fine.