Db,update: Treat second parameter as oldElement
Until the drivers were not unified in whether or not to treat the parameter
as ignorable old value or the new value. That caused issues when storing
properties with the K8sEngine, since it has to ignore the parameter since
it only stores the entire object (all or nothing). PropsContainer, however
used the now "oldElement" for the newValue, which caused issues.