Reconnect: Attempt to fix spurious OtherController msgs
If the controller "loses connection" (i.e. does not get Pong messages
in time) and reconnects, in some cases the satellite might still
have the old connection open.

When the satellite has an open connection to a controller and a new
socket is opened with also a valid 'Auth' packet, the satellite
responds to the old connection 'OtherController'.

If in rare cases (race condition) the old and the new connections
come from the same controller, that controller still receives the
'OtherController' and might close the connection and no longer
tries to reconnect.

This patch attempts to fix this spurious 'OtherController' message
by checking the additionally sent controller's UUID (Cluster/LocalID).
Only if that does not match with the old UUID the 'OtherController'
message is sent.