ToggleDisk: Set resolved StorPoolName property
Otherwise the property is not set at all after a toggle disk operation
when the user omitted the --storage-pool, for example:
  linstor r td node rsc
Linstor will try to check if the given node also has a storage pool
with the same name as its peers and if so will take it. (not ideal,
I guess it would be better to let an autoplacer run limited to that
node).

Not having the property set can cause to confusion for the user
but maybe also to some issues when layer-data needs to be recreated