Automatically fix symlinks for devices
Introduce an additional handler that checks for the device path before each such modification.
If the device is not found, it attempts to fix the symlink using dmsetup output.

This change is workaround for specific set of issues, often related to udev,
which lead to the disappearance of symlinks for LVM devices on a working system.
These issues commonly occur during device resizing and deactivation,
causing LINSTOR expceptions diring accessing DRBD super-block of volume.

fixes: https://github.com/LINBIT/linstor-server/issues/326

Signed-off-by: Andrei Kvapil <kvapss@gmail.com>