DRBD: attempt to (re-) start events2 on "node reconnect"
The broken scenario was the following:
* Have *no* DRBD installed on a node.
* Start the satellite.
  Controller will reject creating DRBD resources on this node
  since it has the information that this node does *not* support
  DRBD9.
* Install DRBD on satellite.
* Do *not* restart satellite but just "node reconnect" it.
  That indeed re-scans for available external tools and
  will let the controller know that it now does support DRBD.
* Try to create a DRBD resource on this satellite.
  The command hangs indefinitely.

The reason for hanging is that since the satellite was started
without DRBD, the `drbdsetup events2` stream did not get started.
Although in the first half of this scenario, not having the events2
stream running was fine since the device-manager waits either until
the DRBD-status is available (from events2) or if DRBD9 is not
supported at all. The second condition allowed the satellite to operate
just fine on non-DRBD resources.

The re-scan during the reconnect changed the supported status, but
did not start the events2 stream. Therefore the second condition was no
longer true (i.e. DRBD is now supported and the condition was that DRBD
is *not* supported), so the device manager had to wait for the first
condition (that means that the DRBD state gets available). However, since
the events2 stream was still not started, this condition will never be
fullfilled (only after a restart of the satellite obviously). That leads
to the device manager to wait/hang indefinitely.