AutoDiskful: Fix Access to deleted Resource
There were 2 ways to reproduce this issue:
Reproducer 1)
Node A:
  * create resource (diskful, only needed so we can create and play
  with a diskless resource on a different node)
Node B:
  * create resource disklessly
  * primary
  * toggle-disk
  * secondary
  * delete
  * primary some other resource

Reproducer 2)
Same as before but instead of actively toggle-disk-ing just wait
the configured AutoDiskful delay and let LINSTOR do the toggle-disk.

The problem with 1) was that the toggle-disk API simply did not clean
up the AutoDiskfulTask entries. The AutoDiskfulTask additionally
ignored (instead of checked for cleanup) diskful resource, so after
the toggle-disk, nothing could remove the entry from AutoDiskfulTask's
internal map.

The fix for this was to obviously notify the AutoDiskfulTask from
the toggle-disk API.

The problem with 2) is a bit trickier, since the AutoDiskfulTask
correctly removed the entry from its list after it triggered the
toggle-disk. However, DRBD apparently sends some new events to the
controller during the toggle-disk process, which re-added the resource
to the AutoDiskfulTask's map where it was just deleted from.

The fix for 1) also fixes this issue since after the toggle-disk is
finished the AutoDiskfulTask is updated again so it can properly cleanup
the resource