Stlt,Auth: Fix version-mismatch check
v1.25.0 introduced a new ctrlUuid in the initial Auth message.
That means that a v1.25.0 satellite will expect that protobuf field
to exist, but a pre v1.25.0 controller will simply not send it.
That leads to an error instead of a reply containing "version mismatch".

The result is that the node stays offline, but the client does not
properly show it as "OFFLINE (Version mismatch)".