sp,l: Fix offline DfltDisklessStorPool
Commit cb538c4647887e19f484b86234fa76c1274c7188 introduced a useful
optimization, but unfortunately introduced a caching problem.

The controller stores possible warnings / issues within a StorPool object
so that every "sp l" can show properly a warning/error if necessary.
A warning as "the node is offline".

The problematic scenario can be reproduced as follows:

* Have a satellite up and running
* Shutdown the satellite
* IMPORTANT: query "sp l"
  This places the "warning"-report within the StorPool object that
  is returned to the client if asked
* Start the satellite and wait until it is online again
* Query "sp l" again

Before this patch, nothing cleared / updated the DfltDisklessStorPool's
reports cache, resulting in the false positive warning that the node
is offline.

This patch fixes this issue by simply including the "fetched free spaces"
from the DfltDisklessStorPool as well (instead of just the thick-SPs as
previously). This goes a bit against the commit mentioned in the beginning,
this this fetching is only done during the FullSync, not during every
"FetchThinFreeSpaces" call.
The trick is, that if the satellite reports the "free space" of the
DfltDisklessStorPool, processing the report also clears the cached
report so that they can be replaced with the new reports (which do not
exist / are null, since the satellite is up and running and online
and everything is fine).