balanceresourcestask: Add task that keeps the resource-group place-count
If diskful resources are missing, it would just call
resource-group adjust on that resource-definition.
If there are too many diskfuls, it would create a list of resources
that aren't allowed to be removed (diskless, inuse, unknown drbd state)
and than use the autoplacer.unplace method to determine the best resources
to remove.