Schedule: fix schedule delete after restore
When a schedule gets deleted, the property
"Schedule/{remoteName}/{scheduleName}/Enabled" also needs to be
removed from all rscDfns that contain the schedule-to-delete.
Since the "Schedule"-namespace also contains other props, and the
scheduleName is accessed by splitting the key into an array, an
IndexOutOfBoundsException ended up occurring. This is now fixed by
simply checking if the property-key has the expected length before
accessing its parts.