Db,migration: added k8s delete crd functionality
This is the first time where LINSTOR needs to make a structural
delete of parts of the database. In this case we are dropping
2 tables, which also affects the CRD version spec.

This rework makes sure that the no longer needed CRDs properly get
deleted from K8s, not just emptied.

This patch also introduces a second constructor for BaseK8sCrdMigration
so that it can perform an automatic check if the structural update
method was called if it was needed.