Schedule: prevent NPE while adding schedule-tasks
If a snapshot that was created via a schedule gets restored, the
resulting rscDfn copies all the props the snapshot had - including
those that tell it to start scheduling. Should a restore like this
happen after the schedule or remote that originally created the
backup was deleted, the resulting ApiCall-message ended up
producing an NPE due to missing null-checks.