ExtCmd: log commands using shell-quoting
In error reports, we sometimes see external commands being used that can't
be copy&pasted because no quotes are used around arguments. This is
particularly present with LVM commands, as the "--config" flag contains
both whitespace and quotes of it's own.

To make this nicer to debug, we escape the logged string if necessary. The
implementation is a port of Pythons "shlex.quote".