docker: remove dmeventd dependency
When running in a container, we cannot assume to have access to dmeventd.
Up to now, we manually started dmeventd to work around an issue when running
the container on systemd with high rlimits. See 0c6c6ec0d1d1 for details.

However, there is no need to start dmeventd at all: it's primary use for
automatic resizing VGs and LVs doesn't fit the use in containers anyways.
So we can just disable "monitoring" of LVs, which in turn should prevent
dmeventd from being started by some random lvcreate.